191일차: 어제 피드백

+++ Study5 - Tree/Week28/Day190/bj10897_wsb.java
@@ -16,15 +16,21 @@ class Main {
         br.close();
     }
 
+    //// 저도 BigInteger 사용않는 로직 시도해보려고, 반복문 한번 돌 때마다 나머지 연산 한번씩 하게했더니 값이 꼬이길래 포기했는데..
+    //// 연산되는 모든 것에 나머지 연산을 먼저 때리고 집어넣으면 제대로 구해지는군요
+    //// 진짜 수학 자신없다 하시지만 누구보다 수학에 진심이신 것 같습니다. 배워갑니다
+    
+    //// 궁금한게 하나 있는데, 나머지를 10억으로 때리니 int형 안에서 충분히 값이 나오지 않나요?
+    //// 혹시 어떤 부분이 int 범위를 초과하는건가요??
     static StringBuilder getCheckListOfJMPABU(int[] checkChildren) {
         StringBuilder sb = new StringBuilder();
 
-        long nodeNum = 0;
+        int nodeNum = 0;
         int generation = 0;
 
-        long nthNode = 1;
-        long prevSiblingNode = 1;
-        long prevLastNodeNum = 0;
+        int nthNode = 1;
+        int prevSiblingNode = 1;
+        int prevLastNodeNum = 0;
 
         for (int checkChild : checkChildren) {
             nthNode = ((nthNode - 1) * ++generation + checkChild) % MOD_NUM;
